requirements.txt
pyyaml>=6.0.1
aiohttp>=3.9.1
python-dotenv>=1.0.0
typing-extensions>=4.9.0
asyncio>=3.4.3
pathlib>=1.0.1 